Update modules names
