+++ src/datasets/categorical.rs
