+++ src/distributions/categorical.rs
